feature hooks y fix pages contact
