+++ src/components/Contact/index.jsx
@@ -1,0 +1,2 @@
+export { ContactForm } from "./ContactForm";
+export { FoxModel } from "./FoxModel";
