+++ src/pages/Contact.jsx
@@ -1,135 +1,35 @@
-import { useAlert } from "../hooks/useAlert";
-import { useRef, useState, Suspense } from "react";
-import emailjs from "@emailjs/browser";
-import { Canvas } from "@react-three/fiber";
-import { Fox } from "../models/Fox/Fox";
-// import Fox from "../models/Fox";
-import { Alert, Loader, Socials } from "../components";
+import { Alert, Socials } from "../components";
+import { ContactForm, FoxModel } from "../components/Contact";
+import { useAlert, useForm, useAnimation, useEmail } from "../hooks";
 
-const Contact = () => {
-  // const formRef = useRef(null)
-  const [form, setForm] = useState({ name: "", email: "", message: "" });
-  const [isLoading, setIsLoading] = useState(false);
-  const [currentAnimation, setCurrentAnimation] = useState("idle");
+export const Contact = () => {
+  const initialState = { name: "", email: "", message: "" };
+  const [form, handleChange, resetForm] = useForm(initialState);
   const { show, text, type, showAlert, hideAlert } = useAlert();
-  const handleChange = (e) => {
-    setForm({ ...form, [e.target.name]: e.target.value });
-  };
-  const handleFocus = () => setCurrentAnimation("walk");
-  const handleBlur = () => setCurrentAnimation("idle");
-  const handleSubmit = (e) => {
-    e.preventDefault();
-    setIsLoading(true);
-    setCurrentAnimation("hit");
-    emailjs
-      .send(
-        import.meta.env.VITE_APP_EMAILJS_SERVICE_ID,
-        import.meta.env.VITE_APP_EMAILJS_TEMPLATE_ID,
-        {
-          from_name: form.name,
-          to_name: "Shaggy",
-          from_email: form.email,
-          to_email: "csoperativo@gmail.com",
-          message: form.message,
-        },
-        import.meta.env.VITE_APP_EMAILJS_PUBLIC_KEY
-      )
-      .then(() => {
-        setIsLoading(false);
-        showAlert("Message sent successfully!", "success");
-        setTimeout(() => {
-          hideAlert();
-          setCurrentAnimation("idle");
-          setForm({ name: "", email: "", message: "" });
-        }, [2000]);
-      })
-      .catch((error) => {
-        setIsLoading(false);
-        setCurrentAnimation("idle");
-        console.log(error);
-        showAlert("Something went wrong!", "error");
-      });
-  };
+  const [currentAnimation, handleFocus, handleBlur, setCurrentAnimation] =
+    useAnimation("idle");
+  const [isLoading, handleSubmit] = useEmail(
+    form,
+    showAlert,
+    hideAlert,
+    resetForm,
+    setCurrentAnimation
+  );
   return (
     <>
       <section className="relative flex lg:flex-row flex-col max-container ">
         {show && <Alert text={text} type={type} />}
-        <div className="flex-1 min-w-[50%] flex flex-col">
-          <h1 className="head-text">Get in Touch</h1>
-          <form
-            className="w-full flex flex-col gap-3 mt-6 "
-            onSubmit={handleSubmit}
-          >
-            <label className="text-black-500 font-semibold">
-              Name
-              <input
-                type="text"
-                name="name"
-                className="input"
-                placeholder="John"
-                required
-                value={form.name}
-                onChange={handleChange}
-                onFocus={handleFocus}
-                onBlur={handleBlur}
-              />
-            </label>
-            <label className="text-black-500 font-semibold">
-              Email
-              <input
-                type="email"
-                name="email"
-                className="input"
-                placeholder="John@gmail.com"
-                required
-                value={form.email}
-                onChange={handleChange}
-                onFocus={handleFocus}
-                onBlur={handleBlur}
-              />
-            </label>
-            <label className="text-black-500 font-semibold">
-              Your Message
-              <textarea
-                name="message"
-                rows={5}
-                className="textarea"
-                placeholder="Let me know I can help you!"
-                // required
-                value={form.message}
-                onChange={handleChange}
-                onFocus={handleFocus}
-                onBlur={handleBlur}
-              />
-            </label>
-            <button
-              type="submit"
-              className="btn"
-              onBlur={handleBlur}
-              onFocus={handleFocus}
-              disabled={isLoading}
-            >
-              {isLoading ? "Sending..." : "Send Message"}
-            </button>
-          </form>
-        </div>
-        <div className="lg:w-1/2 w-full lg:h-auto md:h-[550px] h-[350px]">
-          <Canvas
-            className=""
-            camera={{ position: [0, 0, 5], fov: 75, near: 0.1, far: 1000 }}
-          >
-            <ambientLight intensity={0.5} />
-            <directionalLight position={[0, 0, 1]} intensity={2.5} />
-            <Suspense fallback={<Loader />}>
-              <Fox
-                position={[0.5, 0.35, 0]}
-                rotation={[12.6, -0.6, 0]}
-                scale={[0.5, 0.5, 0.5]}
-                currentAnimation={currentAnimation}
-              />
-            </Suspense>
-          </Canvas>
-        </div>
+        <ContactForm
+          {...{
+            form,
+            isLoading,
+            handleChange,
+            handleFocus,
+            handleBlur,
+            handleSubmit,
+          }}
+        />
+        <FoxModel currentAnimation={currentAnimation} />
       </section>
       <section className="max-social">
         <Socials />
@@ -137,5 +37,3 @@ const Contact = () => {
     </>
   );
 };
-
-export default Contact;
